refactor: 사용되지 않는 Neo4j 전용 레포지토리 제거 및 JPA 레포지토리 디렉토리 정리
